backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.0
email-validator
requests==2.31.0
psycopg2-binary==2.9.9


